vsprintf: make basprintf an alias of xasprintf
Less than 12% of all our basprintf callsites actually bother to check
for allocation failure. It's highly unlikely that barebox will be able
to function when the heap is so full that the usually small basprintf
requests fail.

Therefore, make basprintf an alias for xasprintf, so barebox panics on
OOM instead of possibly triggering undefined behavior.

Users doing big allocations with basprintf and wishing to handle their
failure should switch over to asprintf instead.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.kernel.org/r/20250313064210.74011-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>