soc: ti: k3-navss-ringacc: fix COMPILE_TEST link error
The TI UDMA driver has a fixed dependency on the K3 RINGACC driver,
which is not necessarily satisfied during COMPILE_TEST as the latter
depends on ARCH_K3 being enabled, which is an ARM-only symbol
unavailable when compile testing on other architectures.

Fix this by selecting the code as needed.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.kernel.org/r/20250313073445.764875-3-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>