ddr_spd: fix always true sub-condition
The last subcondition is always true, because it compares the left-hand
side against two different values and ORs the result.

I think the intention of the code is to avoid printing bogus serial
numbers that are either all zeroes or all ones, so let's check for that.

This is a debug print, so there is no fallout expected if the guess
turns out wrong..

Cc: Denis Orlov <denorl2009@gmail.com>
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.kernel.org/r/20250313073445.764875-9-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>