regulator: fan53555: suppress warning about uninitialized variable
In function 'fan53555_device_setup', a call to 'regmap_update_bits' is
made with the mode_mask variable potentially uninitialized.

This can be the case if di->vendor is FAN53526_VENDOR_FAIRCHILD and
di->sleep_vsel_id is not handled by the case (neither FAN53555_VSEL_ID_0 nor FAN53555_VSEL_ID_1).
I am not sure how likely it is to happen.

Signed-off-by: Jules Maselbas <jmaselbas@zdiv.net>
Link: https://lore.kernel.org/r/20250314152437.25675-1-jmaselbas@zdiv.net
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>