amba: support masking data abort during identification
Registration of an AMBA device requires reading its pid and cid
registers after turning on the device. This is the first I/O done to the
device and may end up crashing if the device tree doesn't describe
the correct clock and power domains or if a previous boot stage
should've enabled them, but didn't.

This is not a likely issue, but when it happens it crashes the system
annoyingly early during device tree populating.

Add a symbol that masks the data abort exception around the
identification code to turn the crash into an error message instead.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.kernel.org/r/20250318082507.419785-3-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>