linux/iopoll: fix inverted condition in read_poll_timeout
If there is no timeout, we wait indefinitely and don't care for the
initial time. A fixup to silence a warning because of multiplication in
boolean context added an explicit comparison and this ended up omitting
get_time_ns() in the wrong case.

Flip it to fix, among others, MDIO on RK3568.

Fixes: 554bbc479c09 ("include: linux/iopoll.h: add optional PBL support")
Reported-by: Uwe Kleine-König <uwe@kleine-koenig.org>
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Tested-by: Uwe Kleine-König <uwe@kleine-koenig.org>
Reviewed-by: Uwe Kleine-König <uwe@kleine-koenig.org>
Link: https://lore.barebox.org/20250426081415.3553913-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>